chore: Added dedicated animations page
