+++ src/app.css
@@ -637,6 +637,7 @@ button {
   height: 100vh;
   padding: 50px;
   position: relative;
+  width: 100%;
 }
 
 .landing {
@@ -644,6 +645,7 @@ button {
   flex-direction: column;
   justify-content: center;
   overflow-x: hidden;
+  width: 100%;
   padding: 100px 50px;
 }
 
@@ -652,30 +654,6 @@ button {
   top: 0;
   margin-bottom: 2%;
   width: 95%;
-
-  svg {
-    width: 100%;
-  }
-}
-
-.intro {
-  height: 100vh;
-}
-
-.intro-text {
-  position: relative;
-  height: 100%;
-}
-
-.textL,
-.textR {
-  opacity: 0;
-  position: absolute;
-  top: 50%;
-  transform: translateY(-50%);
-  width: 50%;
-  padding: 0 10%;
-  will-change: transform;
 }
 
 .hold {
@@ -686,7 +664,7 @@ button {
   align-items: center;
 }
 
-.hold .resize {
+.resize {
   width: 100px;
   height: 100px;
   border-radius: 100%;
